requirements.txt
streamlit
pandas
matplotlib.pyplot
altair
time